Refactor all entities
